requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
frozenlist==1.6.0
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
markupsafe==3.0.2
meraki==2.0.2
multidict==6.4.3
packaging==25.0
pluggy==1.5.0
propcache==0.3.1
pyproject-hooks==1.2.0
pytest==7.4.4
python-dotenv==1.1.0
pyyaml==6.0.2
requests==2.32.3
setuptools==70.3.0
tomli==2.2.1
typing-extensions==4.13.2
urllib3==2.4.0
wheel==0.45.1
yarl==1.20.0
